+++ java/pom.xml
@@ -64,10 +64,11 @@
                     <excludes>
                         <exclude>src/main/resources/**</exclude>
                         <exclude>src/test/resources/**</exclude>
-                        <exclude>**/pom.xml</exclude>
+                        <exclude>**/*.xml</exclude>
                         <exclude>**/*.md</exclude>
                         <exclude>**/*.sh</exclude>
                         <exclude>**/*.txt</exclude>
+                        <exclude>**/*.properties</exclude>
                     </excludes>
                 </configuration>
                 <executions>
